requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
bidict==0.21.2
chardet==4.0.0
idna==3.1
multidict==5.1.0
netifaces==0.10.6
python-dotenv==0.17.1
python-engineio==4.1.0
python-socketio==5.2.1
pytz==2021.1
regex==2021.4.4
typing-extensions==3.10.0.0
tzlocal==2.1
watchgod==0.7
yarl==1.6.3
